move lexer into a separate package
