+++ pkg/operations/lexer/scanner/scanner.go
