+++ pkg/operations/lexer/scanner/scanner_test.go
